Add map view into the new morecast page
