+++ web/src/features/fireWeather/components/WxDataForm.tsx
@@ -6,7 +6,6 @@ import TimeOfInterestPicker from 'features/fireWeather/components/TimeOfInterest
 import GetWxDataButton from 'features/fireWeather/components/GetWxDataButton'
 import { stationCodeQueryKey, timeOfInterestQueryKey } from 'utils/url'
 import WxStationDropdown from 'features/fireWeather/components/WxStationDropdown'
-import WeatherMap from 'features/fireWeather/components/maps/WeatherMap'
 
 const useStyles = makeStyles({
   form: {
@@ -84,7 +83,6 @@ const WxDataForm = ({ codesFromQuery, toiFromQuery, openSidePanel }: Props) => {
         stationCodes={selectedCodes}
         onChange={setSelectedCodes}
       />
-      <WeatherMap />
       <TimeOfInterestPicker
         className={classes.toi}
         timeOfInterest={timeOfInterest}
