+++ web/src/features/fireWeather/pages/OldMoreCastPage.tsx
@@ -1,10 +1,10 @@
-import React, { useState, useEffect } from 'react'
+import React, { useEffect } from 'react'
 import { useDispatch } from 'react-redux'
 import { useLocation } from 'react-router-dom'
-import { makeStyles } from '@material-ui/core/styles'
 
-import { PageHeader } from 'components'
-import { getStationCodesFromUrl, getTimeOfInterestFromUrl } from 'utils/url'
+import { PageHeader, PageTitle, Container } from 'components'
+import WxDataDisplays from 'features/fireWeather/components/WxDataDisplays'
+import WxDataForm from 'features/fireWeather/components/WxDataForm'
 import { fetchWxStations } from 'features/percentileCalculator/slices/stationsSlice'
 import { fetchGlobalModelsWithBiasAdj } from 'features/fireWeather/slices/modelsSlice'
 import { fetchObservations } from 'features/fireWeather/slices/observationsSlice'
@@ -15,54 +15,14 @@ import { fetchHighResModels } from 'features/fireWeather/slices/highResModelsSli
 import { fetchHighResModelSummaries } from 'features/fireWeather/slices/highResModelSummariesSlice'
 import { fetchRegionalModels } from 'features/fireWeather/slices/regionalModelsSlice'
 import { fetchRegionalModelSummaries } from 'features/fireWeather/slices/regionalModelSummariesSlice'
-import WxDataDisplays from 'features/fireWeather/components/WxDataDisplays'
-import WxDataForm from 'features/fireWeather/components/WxDataForm'
-import SidePanel from 'features/fireWeather/components/SidePanel'
-import NetworkErrorMessages from 'features/fireWeather/components/NetworkErrorMessages'
-
-const useStyles = makeStyles(theme => ({
-  main: {
-    display: 'flex',
-    flexDirection: 'column',
-    height: '100vh'
-  },
-  nav: {
-    background: theme.palette.primary.light,
-    color: theme.palette.primary.contrastText,
-    minHeight: 60,
-    display: 'flex',
-    flexWrap: 'wrap',
-    alignItems: 'center',
-    paddingLeft: 25,
-    paddingRight: 25
-  },
-  content: {
-    flexGrow: 1,
-    display: 'flex',
-    overflowY: 'auto'
-  },
-  map: {
-    order: 1,
-    flexGrow: 1,
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    fontSize: '2rem'
-  }
-}))
+import { getStationCodesFromUrl, getTimeOfInterestFromUrl } from 'utils/url'
 
 const MoreCastPage = () => {
-  const classes = useStyles()
   const dispatch = useDispatch()
   const location = useLocation()
 
   const codesFromQuery = getStationCodesFromUrl(location.search)
   const toiFromQuery = getTimeOfInterestFromUrl(location.search)
-
-  const shouldInitiallyShowSidePanel = codesFromQuery.length > 0
-  const [showSidePanel, setShowSidePanel] = useState(shouldInitiallyShowSidePanel)
-  const openSidePanel = () => setShowSidePanel(true)
-  const closeSidePanel = () => setShowSidePanel(false)
 
   useEffect(() => {
     dispatch(fetchWxStations())
@@ -83,22 +43,13 @@ const MoreCastPage = () => {
   }, [location]) // eslint-disable-line react-hooks/exhaustive-deps
 
   return (
-    <main className={classes.main}>
-      <PageHeader title="MoreCast" productName="MoreCast" noContainer padding={25} />
-      <div className={classes.nav}>
-        <WxDataForm
-          codesFromQuery={codesFromQuery}
-          toiFromQuery={toiFromQuery}
-          openSidePanel={openSidePanel}
-        />
-      </div>
-      <div className={classes.content}>
-        <div className={classes.map}>map</div>
-        <SidePanel show={showSidePanel} closeSidePanel={closeSidePanel}>
-          <NetworkErrorMessages />
-          <WxDataDisplays stationCodes={codesFromQuery} timeOfInterest={toiFromQuery} />
-        </SidePanel>
-      </div>
+    <main>
+      <PageHeader title="Predictive Services Unit" productName="MoreCast" />
+      <PageTitle title="MoreCast - Weather Forecast Validation Tool" />
+      <Container>
+        {/* <WxDataForm codesFromQuery={codesFromQuery} toiFromQuery={toiFromQuery} /> */}
+        <WxDataDisplays stationCodes={codesFromQuery} timeOfInterest={toiFromQuery} />
+      </Container>
     </main>
   )
 }
